refactor dashboard page
