+++ src/components/DashboardHeader/DashboardHeader.scss
@@ -7,7 +7,7 @@
   justify-content: space-between;
   align-items: center;
   height: rem(50px);
-  width: rem(85%);
+  width: rem(90%);
   padding: rem(15px) rem(30px);
 
   &__hamburger {
