--- src/pages/Dashboard/Dashboard.scss
@@ -1,3 +1,0 @@
-.dashboard {
-  display: flex;
-}
